[Rename] #65 이상한 파일 이름 수정 ColorSchemeModifiers -> GSButtonStyleModifiers
